fix(flake): Setup google-chrome and chromedriver
This commit adds google-chrome and chromedriver to the
systemPackages in flake.nix and sets the corresponding
environment variables in home.nix. This allows using
chrome and chromedriver with tools like Selenium.

The CHROME_BINARY and CHROMEDRIVER_BINARY environment
variables are now available in the fish shell. This
ensures that tools which need these binaries can find
them correctly.